+++ CHANGES.md
@@ -1,2 +1,7 @@
+1.0.0, 08-09-2021
+- Added asynchronous calls for multiple url requests
+  - Config dictionary class
+  - Finding API request builder
+
 0.0.0, 29-08-2021
 - Initial release (setup, no user functional code)